+++ src/model/BasicDeal.js
@@ -211,12 +211,6 @@ BasicDeal['StatusEnum'] = {
     "lost": "lost",
 
     /**
-     * value: "lost"
-     * @const
-     */
-    "lost": "lost",
-
-    /**
      * value: "deleted"
      * @const
      */